Basic implementation, and fix some stuff.
Forgot that custom headers are included with double-quotes, not angled
brackets.
Fix weird thing where I could not allocate a std::mutex statically and
had to do so dynamically, still do not understand why this is the way
things are ... will ask around to figure out what the heck that was
about.

Had to allocate a *pointer* to a mutex, and initialize it in the
constructor ...

Write main.cpp as a starting point.