Changing things around.
Lots of hesitation going on here mostly because i can't reliably predict
what will occur, and, as the last commit somewhat clearly points to the
fact that it's the fault of java, kind of ironic, since i haven't even
used it for that much.

Added a string argument to `Cell`'s constructor, which i'll use when
manually creating the stations in the main.cpp file.

Fix some other minor bugs (like how the mainCPP was trying to access
stations by the cells themselves and not their indices (recall the Cell&
vector 'bug'), and the invalid dereferencing of that associatedTrack
pointer.

I hope some of my intuitions are not completely wrong.