First prototype.
Seems to be working, but the thing is I only tried with one train so I
cant say for sure, but I'm hopeful :+1:

Had to do somme modifications however, I think I'll have to do those
anyways the more I discover things to add.

The ones I added were to the enter() and leave() functions, added a
string argument to them so I can track which train got into which
station/cell.

Kinda pleased it worked.